requirements.txt
spacy
google.genai
dotenv
pytest